requirements.txt
langchain==0.0.146
openai
faiss-cpu
streamlit
streamlit-chat
python-dotenv
tiktoken
